Update signatures and add missing imports

+++ lib/emscripten/src/env.rs
@@ -188,10 +188,10 @@ pub fn ___build_environment(environ: c_int, ctx: &mut Ctx) {
     // };
 }
 
-pub fn _sysconf(name: c_int, _ctx: &mut Ctx) -> c_long {
+pub fn _sysconf(name: c_int, _ctx: &mut Ctx) -> i32 {
     debug!("emscripten::_sysconf {}", name);
     // TODO: Implement like emscripten expects regarding memory/page size
-    unsafe { sysconf(name) }
+    unsafe { sysconf(name) as i32 } // TODO review i64
 }
 
 pub fn ___assert_fail(a: c_int, b: c_int, c: c_int, d: c_int, _ctx: &mut Ctx) {
